Refactor custom features and address code review feedback
- Restored missing LampArray navigation in MainWindow.xaml

- Fixed GPUWatcherService history by passing process updates from GPUController

- Consolidated High-Performance UI into the existing EnableHardwareAcceleration setting

- Removed redundant ProfileSyncService in favor of native backup functionality

- Moved extension structs and classes to Structs.cs for better code organization

- Cleaned up verbose comments in new services
