+++ LenovoLegionToolkit.WPF/Controls/Settings/SettingsAppBehaviorControl.xaml
@@ -95,17 +95,7 @@
                 Icon="Save24">
                 <controls:CardHeaderControl Title="{x:Static resources:Resource.SettingsPage_BackupSettings_Title}" Subtitle="{x:Static resources:Resource.SettingsPage_BackupSettings_Message}" />
             </custom:CardAction>
-            <custom:CardControl Margin="0,0,0,8" Icon="TopSpeed24">
-                <custom:CardControl.Header>
-                    <controls:CardHeaderControl Title="High Performance UI" Subtitle="Enable hardware acceleration and smooth animations (144 FPS). Disable for battery savings. Requires restart." />
-                </custom:CardControl.Header>
-                <wpfui:ToggleSwitch
-                    x:Name="_highPerformanceUIToggle"
-                    Margin="0,0,0,8"
-                    AutomationProperties.Name="High Performance UI"
-                    Click="HighPerformanceUIToggle_Click"
-                    Visibility="Hidden" />
-            </custom:CardControl>
+
             <custom:CardControl Margin="0,0,0,8" Icon="Keyboard24">
                 <custom:CardControl.Header>
                     <controls:CardHeaderControl
